requirements.txt
flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==8.0.28
requests
pandas